Move StdinWriter to process.
